+++ components/movies/EchMoviesCard.vue
@@ -84,17 +84,13 @@ export default {
       }
     }
   },
-  asyncData({ params }) {
-    console.log('params ' + params)
-    return {}
-  },
   data() {
     return {}
   },
   methods: {
     getMovieDetailURL(movie_id, movie_title) {
       const slugger = ServiceLocator.Slugger.sluggify([movie_title])
-      return `/es-ES/movies/${movie_id}/${slugger}`
+      return `/lang/${this.$store.state.language}/movies/${movie_id}/${slugger}`
     },
     getPosterURL(posterPath) {
       return `https://image.tmdb.org/t/p/w185_and_h278_bestv2/${posterPath}`
